source code restructure
