+++ bruno/bruno.json
@@ -1,0 +1,9 @@
+{
+  "version": "1",
+  "name": "bruno",
+  "type": "collection",
+  "ignore": [
+    "node_modules",
+    ".git"
+  ]
+}
