+++ bruno/create data.bru
@@ -1,0 +1,11 @@
+meta {
+  name: create data
+  type: http
+  seq: 2
+}
+
+post {
+  url: http://localhost:6001/api/create
+  body: none
+  auth: inherit
+}
