+++ bruno/root.bru
@@ -1,0 +1,11 @@
+meta {
+  name: root
+  type: http
+  seq: 1
+}
+
+get {
+  url: http://localhost:6001
+  body: none
+  auth: inherit
+}
