+++ docker-compose.yml
@@ -1,0 +1,7 @@
+services:
+  mongo:
+    image: mongo:latest
+    volumes:
+      - './data/db:/data/db'
+    ports:
+      - 27017:27017
