+++ src/config/cors.config.js
